refactor: 공용 로직 패키지 구조 수정
