+++ src/main/java/com/example/daobe/common/presentation/cookie/CookieHandler.java
@@ -1,4 +1,4 @@
-package com.example.daobe.common.controller.cookie;
+package com.example.daobe.common.presentation.cookie;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseCookie;
